requirements.txt
asgiref==3.2.3
astroid==2.3.3
Django==3.0.1
django-crispy-forms==1.8.1
html5lib==1.0.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==6.2.1
pylint==2.4.4
PyPDF2==1.26.0
pytz==2019.3
reportlab==3.5.32
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
webencodings==0.5.1
wrapt==1.11.2
xhtml2pdf==0.2.3
